tenshi.openrc-init: remove redundant PID file checks.
The "reload" and "flush" commands in the OpenRC init script were
performing a check to ensure that tenshi was running before they did
anything. However, since those two commands are defined to be
"extra_started_commands", they will only be available when tenshi is
"started", i.e. running. Thus the manual check for the PID file is
redundant.